陽性者数（区市町村別）: Rename components/cards/ConfirmedCasesByMunicipalitiesCard.vue to components/index/CardsReference/ConfirmedCasesByMunicipalities/Card.vue
